move multikernelmanager to IPython.kernel
